backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# GCP
google-cloud-storage==2.18.2
google-cloud-documentai==2.32.0
google-cloud-firestore==2.19.0
google-cloud-tasks==2.17.1

# PDF processing
pypdf==5.1.0
pdf2image==1.17.0
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
